update projects to make them compilable

+++ foodauthent-osgi/foodauthent-api-internal/src/main/java/org/foodauthent/internal/api/persistence/PersistenceService.java
@@ -102,12 +102,12 @@ public interface PersistenceService {
 	/**
 	 * @return number of pages available in total
 	 */
-	long getTotalNumPages();
+	int getTotalNumPages();
 	
 	/**
 	 * @return total number of entries available
 	 */
-	long getTotalNumEntries();
+	int getTotalNumEntries();
 	
 	/**
 	 * @return the actual result list
